+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - New `hdfs` input and output.
+- New `switch` output.
 
 ### 0.29.0 - 2018-09-02